+++ hikyuu_cpp/hikyuu/trade_sys/moneymanager/imp/FixedCountMoneyManager.cpp
@@ -10,19 +10,19 @@
 namespace hku {
 
 FixedCountMoneyManager::FixedCountMoneyManager() : MoneyManagerBase("MM_FixedCount") {
-    setParam<int>("n", 100);
+    setParam<double>("n", 100);
 }
 
 FixedCountMoneyManager::~FixedCountMoneyManager() {}
 
-size_t FixedCountMoneyManager ::_getBuyNumber(const Datetime& datetime, const Stock& stock,
+double FixedCountMoneyManager ::_getBuyNumber(const Datetime& datetime, const Stock& stock,
                                               price_t price, price_t risk, SystemPart from) {
-    return getParam<int>("n");
+    return getParam<double>("n");
 }
 
-MoneyManagerPtr HKU_API MM_FixedCount(int n) {
+MoneyManagerPtr HKU_API MM_FixedCount(double n) {
     FixedCountMoneyManager* p = new FixedCountMoneyManager();
-    p->setParam<int>("n", n);
+    p->setParam<double>("n", n);
     return MoneyManagerPtr(p);
 }
