+++ hikyuu_cpp/hikyuu/trade_sys/moneymanager/imp/FixedRatioMoneyManager.h
@@ -20,7 +20,7 @@ public:
 
     virtual void _reset() override;
     virtual MoneyManagerPtr _clone() override;
-    virtual size_t _getBuyNumber(const Datetime& datetime, const Stock& stock, price_t price,
+    virtual double _getBuyNumber(const Datetime& datetime, const Stock& stock, price_t price,
                                  price_t risk, SystemPart from) override;
 
 private:
